26 Manually binding into Laravel's service container/IOC container
Useful, if there’s ever a point in a various class you need access to
some object, some dependency, but the problem is that dependency has to
be constructed in a specific way. You have to give it an API key, you
have to give it a dependency. You don’t want to prepare that
instantiation over and over. Do it once and bind it once into Laravel's
service container/IOC container, the same thing. That way, whenever you
need it, for classes that are resolved outside of the service
container, you just type hint what you want and Laravel will give you
what you specified.