26 constructor injection
Laravel is looking into it’s service container to give you what you
require. In this case, we didn’t bind anything, but it saw that they
want FooRepository. We can instantiate that object on our own, so I am
going to do it for the user and do it in the constructor. Sometimes
you’ll have a controller method, where you need some kind of
dependency, but sometimes that method is where you need the dependency.